+++ Domain/PurchaseDetail.cs
@@ -17,7 +17,7 @@ namespace Domain
         public int Id { get; set; }
         public int PurchaseId { get; set; }
         public int ProductId { get; set; }
-        public int ProductValue { get; set; }
+        public decimal ProductValue { get; set; }
         public int Quantity { get; set; }
     
         public virtual Product Product { get; set; }
